update sodium to the newest version replace overwrites with injects (workaround to fix build task)
