+++ .github/workflows/build.yml
@@ -22,24 +22,24 @@ jobs:
         with:
           java-version: ${{ matrix.java }}
 
-      - name: Gradle Build Forge
+      - name: Gradle Build NeoForge
         uses: gradle/gradle-build-action@v2.4.2
         with:
-          arguments: Forge:jar
+          arguments: neoforge:jar
 
       - name: Capture Build Artifacts Forge
         uses: actions/upload-artifact@v3
         with:
-          name: Artifacts Forge
-          path: Forge/build/libs/*.jar
+          name: Artifacts NeoForge
+          path: neoforge/build/libs/*.jar
 
       - name: Gradle Build Fabric
         uses: gradle/gradle-build-action@v2.4.2
         with:
-          arguments: Fabric:build
+          arguments: fabric:build
 
       - name: Capture Build Artifacts Fabric
         uses: actions/upload-artifact@v3
         with:
           name: Artifacts Fabric
-          path: Fabric/build/libs/*.jar
+          path: fabric/build/libs/*.jar
