+++ .idea/inspectionProfiles/Project_Default.xml
@@ -3,5 +3,6 @@
     <option name="myName" value="Project Default" />
     <inspection_tool class="BooleanMethodIsAlwaysInverted" enabled="false" level="WARNING" enabled_by_default="false" />
     <inspection_tool class="UnqualifiedFieldAccess" enabled="true" level="WARNING" enabled_by_default="true" />
+    <inspection_tool class="UnusedMixin" enabled="false" level="WARNING" enabled_by_default="false" />
   </profile>
 </component>
