+++ fabric/build.gradle
@@ -20,7 +20,7 @@ dependencies {
     
     // For compatibility
 	modImplementation "maven.modrinth:sodium:${sodium_version}-fabric"
-	modImplementation "maven.modrinth:lithium:${lithium_version}"
+	modImplementation "maven.modrinth:lithium:${lithium_version}-fabric"
 
     compileOnly "org.checkerframework:checker-qual:3.12.0"
     compileOnly group: 'org.projectlombok', name: 'lombok', version: "${lombok_version}"
