+++ fabric/src/main/java/de/leximon/fluidlogged/mixin/classes/compat/lithium/LevelChunkSectionMixin.java
@@ -1,13 +1,7 @@
-package de.leximon.fluidlogged.mixin.classes.fabric.compat_lithium;
+package de.leximon.fluidlogged.mixin.classes.compat.lithium;
 
-import de.leximon.fluidlogged.mixin.extensions.LevelChunkSectionExtension;
-import me.jellysquid.mods.lithium.common.block.BlockStateFlags;
-import me.jellysquid.mods.lithium.common.block.TrackedBlockStatePredicate;
 import net.minecraft.world.level.chunk.LevelChunkSection;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(LevelChunkSection.class)
 public class LevelChunkSectionMixin {
