+++ fabric/src/main/java/de/leximon/fluidlogged/mixin/classes/compat/sodium/ClonedChunkSectionMixin.java
@@ -1,4 +1,4 @@
-package de.leximon.fluidlogged.mixin.classes.fabric.compat_sodium;
+package de.leximon.fluidlogged.mixin.classes.compat.sodium;
 
 import de.leximon.fluidlogged.mixin.extensions.LevelChunkSectionExtension;
 import de.leximon.fluidlogged.mixin.extensions.compat_sodium.ClonedChunkSectionExtension;
