--- fabric/src/main/java/de/leximon/fluidlogged/mixin/classes/fabric/compat_sodium/ChunkBuilderMeshingTaskMixin.java
@@ -1,67 +1,0 @@
-package de.leximon.fluidlogged.mixin.classes.fabric.compat_sodium;
-
-import de.leximon.fluidlogged.mixin.extensions.compat_sodium.WorldSliceExtension;
-import net.caffeinemc.mods.sodium.client.world.LevelSlice;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.material.FluidState;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
-import org.spongepowered.asm.mixin.injection.Redirect;
-
-@Mixin(targets = "net/caffeinemc/mods/sodium/client/render/chunk/compile/tasks/ChunkBuilderMeshingTask")
-public class ChunkBuilderMeshingTaskMixin {
-
-    @Unique private FluidState fluidlogged$fluidState;
-    
-    @Redirect(
-      method = "execute(Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildContext;Lnet/caffeinemc/mods/sodium/client/util/task/CancellationToken;)Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildOutput;",
-      at = @At(
-        value = "INVOKE",
-        target = "Lnet/caffeinemc/mods/sodium/client/world/LevelSlice;getBlockState(III)Lnet/minecraft/world/level/block/state/BlockState;",
-        ordinal = 0
-      )
-    )
-    private BlockState redirectCaptureFluidStateFabric(LevelSlice instance, int x, int y, int z) {
-        this.fluidlogged$fluidState = ((WorldSliceExtension) (Object) instance).fluidlogged$getFluidState(x, y, z);
-        return instance.getBlockState(x, y, z);
-    }
-    
-    @Redirect(
-      method = "execute(Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildContext;Lnet/caffeinemc/mods/sodium/client/util/task/CancellationToken;)Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildOutput;",
-      at = @At(
-        value = "INVOKE",
-        target = "Lnet/minecraft/world/level/block/state/BlockState;isAir()Z",
-        ordinal = 0
-      )
-    )
-    private boolean redirectIsAir(BlockState instance) {
-        return instance.isAir() && this.fluidlogged$fluidState.isEmpty();
-    }
-    
-    @Redirect(
-      method = "execute(Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildContext;Lnet/caffeinemc/mods/sodium/client/util/task/CancellationToken;)Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildOutput;",
-      at = @At(
-        value = "INVOKE",
-        target = "Lnet/minecraft/world/level/material/FluidState;isEmpty()Z"
-      )
-    )
-    private boolean redirectIsEmpty(FluidState instance) {
-        return this.fluidlogged$fluidState.isEmpty();
-    }
-    
-    @ModifyArg(
-      method = "execute(Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildContext;Lnet/caffeinemc/mods/sodium/client/util/task/CancellationToken;)Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildOutput;",
-      at = @At(
-        value = "INVOKE",
-        target = "Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/pipeline/FluidRenderer;render(Lnet/caffeinemc/mods/sodium/client/world/LevelSlice;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/material/FluidState;Lnet/minecraft/core/BlockPos;Lnet/minecraft/core/BlockPos;Lnet/caffeinemc/mods/sodium/client/render/chunk/translucent_sorting/TranslucentGeometryCollector;Lnet/caffeinemc/mods/sodium/client/render/chunk/compile/ChunkBuildBuffers;)V"
-      ),
-      index = 2,
-      remap = false
-    )
-    private FluidState modifyPassedFluidState(FluidState fluidState) {
-        return fluidState.isEmpty() ? this.fluidlogged$fluidState : fluidState;
-    }
-
-}
