+++ fabric/src/main/resources/fluidlogged.fabric.mixins.json
@@ -5,13 +5,13 @@
   "compatibilityLevel": "JAVA_21",
   "plugin": "de.leximon.fluidlogged.mixin.FluidloggedFabricMixinPlugin",
   "mixins": [
-    "fabric.compat_sodium.ChunkBuilderMeshingTaskMixin",
-    "fabric.compat_sodium.ClonedChunkSectionMixin",
-    "fabric.compat_sodium.FluidRendererMixin",
-    "fabric.compat_sodium.WorldSliceMixin",
-    "fabric.compat_lithium.LevelChunkSectionMixin",
-    "fabric.compat_milk_lib.MilkBucketItemMixin",
-    "fabric.world_interaction.removal_and_placement.BucketItemMixin",
+    "removal_and_placement.BucketItemMixin",
+    "compat.sodium.BlockOcclusionCacheMixin",
+    "compat.sodium.ChunkBuilderMeshingTaskMixin",
+    "compat.sodium.ClonedChunkSectionMixin",
+    "compat.sodium.FluidRendererMixin",
+    "compat.sodium.WorldSliceMixin",
+    "compat.lithium.LevelChunkSectionMixin",
     "network.ChunkHolderMixin",
     "network.ServerChunkCacheMixin",
     "network.ServerLevelMixin",
