+++ gradle.properties
@@ -48,5 +48,5 @@ debug_publish=true
 # Dependencies
 modmenu_version=11.0.0
 yacl_version=3.5.0+1.21
-sodium_version=mc1.21-0.6.0-beta.2
-lithium_version=mc1.21.1-0.13.0
+sodium_version=mc1.21.1-0.6.5
+lithium_version=mc1.21.1-0.14.3
