+++ neoforge/src/main/resources/fluidlogged.neoforge.mixins.json
@@ -5,11 +5,13 @@
   "compatibilityLevel": "JAVA_21",
   "plugin": "de.leximon.fluidlogged.mixin.FluidloggedForgeMixinPlugin",
   "mixins": [
-    "neoforge.compat_sodium.ChunkBuilderMeshingTaskMixin",
-    "neoforge.compat_sodium.ClonedChunkSectionMixin",
-    "neoforge.compat_sodium.FluidRendererMixin",
-    "neoforge.compat_sodium.WorldSliceMixin",
-    "neoforge.world_interaction.removal_and_placement.BucketItemMixin",
+    "world_interaction.removal_and_placement.BucketItemMixin",
+    "compat.sodium.BlockOcclusionCacheMixin",
+    "compat.sodium.ChunkBuilderMeshingTaskMixin",
+    "compat.sodium.ClonedChunkSectionMixin",
+    "compat.sodium.FluidRendererMixin",
+    "compat.sodium.WorldSliceMixin",
+    "compat.lithium.LevelChunkSectionMixin",
     "network.ChunkHolderMixin",
     "network.ServerChunkCacheMixin",
     "network.ServerLevelMixin",
